third_party/WebKit/LayoutTests/inspector/console/console-dir-es6-expected.txt

CONSOLE MESSAGE: line 15: [object Object]
CONSOLE MESSAGE: line 18: Symbol()
CONSOLE MESSAGE: line 22: [object Map]
CONSOLE MESSAGE: line 22: [object WeakMap]
CONSOLE MESSAGE: line 26: [object Set]
CONSOLE MESSAGE: line 26: [object WeakSet]
Tests that console logging dumps proper messages.

console-dir-es6.html:15 Object
    a: 1
    Symbol(): 2
    Symbol(Symbol.iterator): Symbol(foo)
    Symbol(a): 3
    Symbol(a): Symbol(Symbol.iterator)
    __proto__: Object
console-dir-es6.html:18 Symbol()
console-dir-es6.html:22 Map
    size: (...)
    __proto__: Map
    <entries>[1]
    0: {key:Object, value:Object}
console-dir-es6.html:22 WeakMap
    __proto__: WeakMap
    <entries>[1]
    0: {key:Object, value:Object}
console-dir-es6.html:26 Set
    size: (...)
    __proto__: Set
    <entries>[1]
    0: Object
console-dir-es6.html:26 WeakSet
    __proto__: WeakSet
    <entries>[1]
    0: Object
